Adding a main util module that allows us to use the validator from other modules.
Adding a main util module that allows us to use the validator from other modules. Also removing status spinner and node-status, since it does not play well with console.log.
