+++ package.json
@@ -22,7 +22,7 @@
     "validate-office-addin": "./app/index.js"
   },
   "author": "OfficeDev",
-  "main": "",
+  "main": "./app/util.js",
   "keywords": [
     "office",
     "validation",
@@ -41,7 +41,6 @@
     "applicationinsights": "^0.19.0",
     "chalk": "^1.1.3",
     "commander": "^2.9.0",
-    "node-status": "^1.0.0",
     "request": "^2.81.0",
     "request-promise": "^4.2.0"
   },
